todo/requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
djangorestframework==3.15.2
pytest==8.3.5
pytest-django==4.9.0
django-cors-headers==4.6.0
